Rename schedule/conf/setting.py for a flatter code layout
The older layout added little to no advantage.
